start porting dynamic value stuff
