+++ subxt/src/custom_values/custom_value_address.rs
@@ -2,7 +2,7 @@ use derivative::Derivative;
 use std::marker::PhantomData;
 
 use crate::dynamic::DecodedValueThunk;
-use crate::metadata::DecodeWithMetadata;
+use subxt_core::metadata::DecodeWithMetadata;
 
 /// This represents the address of a custom value in in the metadata.
 /// Anything, that implements the [CustomValueAddress] trait can be used, to fetch
