+++ subxt/src/storage/utils.rs
@@ -7,7 +7,9 @@
 //! the trait itself.
 
 use super::StorageAddress;
-use crate::{error::Error, metadata::Metadata};
+use crate::error::Error;
+
+use subxt_core::metadata::Metadata;
 
 /// Return the root of a given [`StorageAddress`]: hash the pallet name and entry name
 /// and append those bytes to the output.
